# Optimize license-check performance