This reduces execution time of license-check and makes it insignificant

| | Before | After |
| -- | -- | -- |
| license-check execution time | 750ms | 4ms |